Working on a file cache.

+++ src/main/java/llc/ufwa/data/resource/cache/FileSegment.java
@@ -2,21 +2,21 @@ package llc.ufwa.data.resource.cache;
 
 public class FileSegment {
     
-    private final long length;
-    private final long index;
+    private final int length;
+    private final int index;
 
-    public FileSegment(final long length, final long index) {
+    public FileSegment(final int length, final int index) {
         
         this.length = length; 
         this.index = index;
         
     }
 
-    public long getLength() {
+    public int getLength() {
         return length;
     }
 
-    public long getIndex() {
+    public int getIndex() {
         return index;
     }
